Add exclude_commits feature
commits can now be ignored using the `git_exclude_commits` config key.
This key is defined with a list of hashes strings to ignore.
This list is then converted into a set of UTF-8 encoded binary objects
to optimize the lookup.
If a commit matches this set, then it is skipped. This can lead to
unhandled files, thus the assert is converted into a warning.