Support "source-read" being emitted for included docs
Since Sphinx 7.2, the "source-read" event is emitted for included
documents (.. include directive). A document can be included multiple
times, and thus the hook cannot assume that it's called only once per
docname.

An issue in Sphinx itself affects the extension when using included
documents. Add a warning in the README.